project/requirements.txt
amqp==1.4.9
anyjson==0.3.3
Babel==2.6.0
billiard==3.3.0.23
celery==3.1.26.post2
diff-match-patch==20121119
Django==1.11.4
django-celery==3.2.2
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.0.1
django-redis==4.9.0
django-reversion==3.0.0
django-tinymce==2.7.0
et-xmlfile==1.0.1
flower==0.9.2
future==0.16.0
gunicorn==19.9.0
httplib2==0.11.3
jdcal==1.4
kombu==3.0.37
odfpy==1.3.6
openpyxl==2.5.4
Pillow==5.2.0
pytz==2018.5
PyYAML==3.13
redis==2.10.6
six==1.11.0
tablib==0.12.1
tornado==5.1
unicodecsv==0.14.1
xlrd==1.1.0
xlwt==1.3.0
Markdown==2.6.11
